Update docs for the documented NHC PeakMode action
