+++ custom_components/nhc2/binary_sensor.py
@@ -43,7 +43,6 @@ from .entities.naso_smartplug_switching_only import Nhc2NasoSmartplugSwitchingOn
 from .entities.naso_smartplug_report_instant_usage import Nhc2NasoSmartplugReportInstantUsageEntity
 from .entities.overallcomfort_action_start_active import Nhc2OverallcomfortActionStartActiveEntity
 from .entities.overallcomfort_action_all_started import Nhc2OverallcomfortActionAllStartedEntity
-from .entities.peakmode_action_basicstate import Nhc2PeakmodeActionBasicStateEntity
 from .entities.playerstatus_action_basicstate import Nhc2PlayerstatusActionBasicStateEntity
 from .entities.timeschedule_action_active import Nhc2TimeschedulActionActiveEntity
 from .nhccoco.devices.accesscontrol_action import CocoAccesscontrolAction
