--- custom_components/nhc2/entities/peakmode_action_button.py
@@ -1,19 +1,0 @@
-from homeassistant.components.button import ButtonEntity
-
-from ..nhccoco.devices.peakmode_action import CocoPeakmodeAction
-from .nhc_entity import NHCBaseEntity
-
-
-class Nhc2PeakmodeActionButtonEntity(NHCBaseEntity, ButtonEntity):
-    _attr_has_entity_name = True
-    _attr_name = None
-
-    def __init__(self, device_instance: CocoPeakmodeAction, hub, gateway):
-        """Initialize a button."""
-        super().__init__(device_instance, hub, gateway)
-
-        self._attr_unique_id = device_instance.uuid
-
-    async def async_press(self) -> None:
-        self._device.press(self._gateway)
-        self.schedule_update_ha_state()
